pyproject.toml
[project]
name = "airflow-udacity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9, <3.13"
dependencies = [
    "apache-airflow[amazon]==3.0.2",
    "apache-airflow-providers-amazon>=9.0.0",
    "apache-airflow-providers-postgres>=6.0.0",
]
